refactor(commands): mod update_notifications
Split everything related to update notifications out to a separate
module because `commands` is getting very big and hard to navigate.
